Added IROS strategy synthesis code. Added code for game construction that is compatable with this str synthesis algorithm.
main.py : Added iros solver method and flag to call that method. get_iros_ts attribute in MiniGridGrpah calls the appropriate build function in minigrid class and also in the product graph class.

iros_solver.py : module that implements the iros str synthesis code given a game that aligns with the abstraction in the paper. To build this particular abstraction set _iros_ts to True in MiniGridGraph class.

product.py : added  _construct_prod_from_iros_ts() that extends the previous absorbing build method with transitions that are assigned players. This required for the str synthesis to work.