Bug Fix: infinite payoff. Regret str when there are more than one way to achieve the given task
infinite_payoff.py: added code to hash a given play. This code previously was committed in non-working condition. NOTE TO SELF: DO NOT COMMIT CODE THAT DOES NOT WORK.

main.py: Changed the formula to reflect in task completion.

NOTE: For finite payoff. we can have some heuristics associated with DFA states. Need to experiment with this