Added finite Val computation.
Added implementation for finite Val computation. Currently I add a self loop to all the states in the product graph(G). The weights of these self-loops are computed using the str_map dictionary in dfa.py module. The cost associated with each state in DFA kind of reflects "reward" to complete certain degree of task(as per the task automaton - DFA).

If you stay at init state the weight of the self loop is 0, if you proceed one step towards the task completion, you get some finite reward. So weights are assigned in increasing order with accepting state with the highest possible +ve weight. I did not stick with negative weights as we can find a loop within G that gives a higher payoff than ending in a state that indicating certain degree of task completion.

In product grpah we look at the dfa node of a given product node tuple. As the current product construction method looks at the current state of the TS this leads to edge weights not being assigned correctly to self loops. Say I am in the state where I observed water(50% task completion), the edge weight should 1 instesd of 0. Due to the current product construction, the dfa at the node where we do observe water is still T0_init. Only after we tak a transition, and are done with the human transition, we update the dfa to reflect what we observed.To avoid this, I look 2 steps forward while adding the edge weights of the self loop.

NOTE: In minigrid, I do not add self loops in raw_trans_sys aka minigrid_TS. This due to the fact that, the system is incentivized to take this self-loop and do nothing. This self-loop my concept of using self-loop as virtual stopping criteria for an infinite game.

dfa.py: Added str_map computation. This implementation is similar to the iros_solver's solve method - with negative weights.

product.py: Added code for adding the so called self-loops as mentioned above. NOTE: I do not add self-loop to the initial state(s), absorbing state(s) i.e accepting state(s) and trap state(s) and states that only transit to the absorbing state(s). I also do not add these self-transition to human nodes(duh!). All methods are trigged when the finite flag is set true.

iros_solver: added some comments

minigrid.py: Added some lines to ignore self-loops

regret_str_synthesis: control str computation - as I can have eve to eve transition, added code to accomodate it.