Added cumulative payoff implememtation for two player games
NOTE: I do not add self-loops from the minigrid TS for the raw_minifris_TS. This has significant effect regret startegy computed. We might have to remove this in future.

cumulative_payoff.py: Following the dynamic programming approach in the iros solver, implemented a similar algorithm for explicit two player game in a cooperative setting. Also added method to visualize the corresponding strategy along with the best cost to reach the accepting state in cooperative setting. The implementation look right as of right! Need to test on different graphs.

minigrid.py: Added if condition to ignore self-loops

main.py: Added call to the newly added cumulative payoff class and solver with corresponding flags.