Refacotered code from recursion to iterative DFS to detect and purge nodes that are not reacahable from the init state of a game
NOTE: Currently, I am suffering from state explosion problem. A simple game with 35 states, with max weight of 34, blows up to ~84,000 states in the graph of utilities which is of the form of a target weighted arena. Each state in this TWA is duplicated by the number of unique target states(|besta alternatives from each edge|) + 1, in our case the states then double to 168,000 states. I don't know how to deal with this as of right now.