Added code to construct camera and sun transition system for ther apriori imaging survey problem. Added plotter class to help plot the game
src/graph/trans_sys.py: Updated class with classmethods to construct the TS for a camera and the sun problem. Updated the product to accomodate for empty aps for the sun transition system.

src/spot/spot.py: removed the "--complete" flag to reduce the size of the DFA by ignoring the edges to the absorbing state.

main.py: UPdated the multitool code and added a plotter class for ease of plotting a dynamic game.